requirements.txt
numpy
fastapi
uvicorn
tensorflow
Pillow
gunicorn
python-multipart